[frontend] Rename comparison graphs
